add softDelete on order state

+++ src/repositories/OrderStateRepository.php
@@ -155,7 +155,13 @@ class OrderStateRepository extends Repository
         foreach ($payplug_os_id_list as $payplug_os_id) {
             if (!in_array($payplug_os_id, $used_os_id_list) && !in_array($payplug_os_id, $used_order_os_id_list)) {
                 $os = new OrderStateEntity($payplug_os_id);
-                $os->delete();
+
+                // Use ObjectModel::softDelete from 1.7.7.0
+                if (version_compare(_PS_VERSION_, '1.7.6', '>')) {
+                    $os->softDelete();
+                } else {
+                    $os->delete();
+                }
             }
         }
     }
